requirements.txt
streamlit
anthropic
langchain
langchain-anthropic
langchain-community
langchain-text-splitter
faiss-cpu
pypdf
python-dotenv
sentence-transformers
google-generativeai
google-genai
rank_bm25
pytz
weasyprint